Refactor Errors and add global configuration iface
