+++ lib/paseto.rb
@@ -9,6 +9,7 @@ require 'rbnacl'
 require 'securerandom'
 require 'sorbet-runtime'
 
+require_relative 'paseto/errors'
 require_relative 'paseto/sodium/stream/base'
 require_relative 'paseto/sodium/stream/x_cha_cha20_xor'
 
@@ -27,42 +28,8 @@ require_relative 'paseto/v4/local'
 require_relative 'paseto/v4/public'
 
 module Paseto
-  # Generic superclass for all Paseto errors
-  class Error < StandardError; end
-
-  # Deserialized data did not include mandatory fields.
-  class ParseError < Error; end
-
-  # Superclass for claim validation errors
-  class ValidationError < Error; end
-
-  # Token is expired
-  class ExpiredToken < ValidationError; end
-
-  # Token has a nbf before the current time
-  class InactiveToken < ValidationError; end
-
-  # Disallowed issuer
-  class InvalidIssuer < ValidationError; end
-
-  # Incorrect audience
-  class InvalidAudience < ValidationError; end
-
-  # Token issued in the future
-  class FutureTokenError < ValidationError; end
-
-  # A cryptographic primitive has failed for any reason,
-  # such as attempting to initialize a stream cipher with
-  # an invalid nonce.
-  class CryptoError < Error; end
-
-  # An authenticator was forged or otherwise corrupt
-  class InvalidAuthenticator < CryptoError; end
-
-  # A signature was forged or otherwise corrupt
-  class InvalidSignature < CryptoError; end
-
   extend T::Sig
+  extend Configuration
 
   include Version
 end
